+++ app/src/androidTest/java/woowacourse/movie/view/activity/MainActivityTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.view.activity
 
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
@@ -10,7 +10,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
-import woowacourse.movie.view.activity.MainActivity
+import woowacourse.movie.R
 
 @RunWith(AndroidJUnit4::class)
 class MainActivityTest {
